Extract repository-related models to their own modules
